add print to helpers.js and surpress console.logs
